CLN: Started to split code
